fix: css 변경 및 목 생성
