+++ src/components/Carousel/index.ts
@@ -1,2 +1,3 @@
 export { default } from './Carousel';
 export * from './type';
+export { default as pendingData } from './mock';
